Move {lib/policy => lib/policies}/*_policy.js
