+++ lib/policies/apply_policy.js
@@ -19,23 +19,46 @@
 const BasePolicy = require('./base_policy')
 
 /**
- * A policy affecting the behavior of scan operations.
+ * A policy affecting the behavior of apply operations.
  *
- * @extends BasePolicy
- * @see {@link module:aerospike.policy} for supported policy values.
+ * @extends ApplyPolicy
  */
-class ScanPolicy extends BasePolicy {
+class ApplyPolicy extends BasePolicy {
   constructor (props) {
     props = props || {}
     super(props)
 
     /**
-     * Abort the scan if the cluster is not in a stable state.
+     * Specifies the behavior for the generation value.
      *
-     * @type boolean
-     * @default false
+     * @type number
+     * @see {@link module:aerospike/policy.gen} for supported policy values.
      */
-    this.failOnClusterChange = props.failOnClusterChange
+    this.gen = props.gen
+
+    /**
+     * Specifies the behavior for the key.
+     *
+     * @type number
+     * @see {@link module:aerospike/policy.key} for supported policy values.
+     */
+    this.key = props.key
+
+    /**
+     * Specifies the number of replicas required to be committed successfully
+     * when writing before returning transaction succeeded.
+     *
+     * @type number
+     * @see {@link module:aerospike/policy.commitLevel} for supported policy values.
+     */
+    this.commitLevel = props.commitLevel
+
+    /**
+     * The time-to-live (expiration) of the record in seconds.
+     *
+     * @type number
+     */
+    this.ttl = props.ttl
 
     /**
      * Specifies whether a {@link
@@ -44,10 +67,10 @@ class ScanPolicy extends BasePolicy {
      * this operation.
      *
      * @type boolean
-     * @default false (do not tombstone deleted records)
+     * @default <code>false</code> (do not tombstone deleted records)
      */
     this.durableDelete = props.durableDelete
   }
 }
 
-module.exports = ScanPolicy
+module.exports = ApplyPolicy
