+++ lib/policies/base_policy.js
@@ -29,13 +29,17 @@ class BasePolicy {
     /**
      * Socket idle timeout in milliseconds when processing a database command.
      *
-     * If socketTimeout is not zero and the socket has been idle for at least socketTimeout,
-     * both maxRetries and totalTimeout are checked. If maxRetries and totalTimeout are not
-     * exceeded, the transaction is retried.
+     * If <code>socketTimeout</code> is not zero and the socket has been idle
+     * for at least <code>socketTimeout</code>, both <code>maxRetries</code>
+     * and <code>totalTimeout</code> are checked. If <code>maxRetries</code>
+     * and <code>totalTimeout</code> are not exceeded, the transaction is
+     * retried.
      *
-     * If both socketTimeout and totalTimeout are non-zero and socketTimeout > totalTimeout,
-     * then socketTimeout will be set to totalTimeout.  If socketTimeout is zero, there will be
-     * no socket idle limit.
+     * If both <code>socketTimeout</code> and <code>totalTimeout</code> are
+     * non-zero and <code>socketTimeout</code> > <code>totalTimeout</code>,
+     * then <code>socketTimeout</code> will be set to
+     * <code>totalTimeout</code>. If <code>socketTimeout</code> is zero, there
+     * will be no socket idle limit.
      *
      * @type number
      * @default 0 (no socket idle time limit).
@@ -45,13 +49,15 @@ class BasePolicy {
     /**
      * Total transaction timeout in milliseconds.
      *
-     * The totalTimeout is tracked on the client and sent to the server along with
-     * the transaction in the wire protocol. The client will most likely timeout
-     * first, but the server also has the capability to timeout the transaction.
+     * The <code>totalTimeout</code> is tracked on the client and sent to the
+     * server along with the transaction in the wire protocol. The client will
+     * most likely timeout first, but the server also has the capability to
+     * timeout the transaction.
      *
-     * If totalTimeout is not zero and totalTimeout is reached before the transaction
-     * completes, the transaction will return error AEROSPIKE_ERR_TIMEOUT.
-     * If totalTimeout is zero, there will be no total time limit.
+     * If <code>totalTimeout</code> is not zero and <code>totalTimeout</code>
+     * is reached before the transaction completes, the transaction will return
+     * error {@link module:aerospike/status.ERR_TIMEOUT|ERR_TIMEOUT}.
+     * If <code>totalTimeout</code> is zero, there will be no total time limit.
      *
      * @type number
      * @default 1000
@@ -62,13 +68,14 @@ class BasePolicy {
      * Maximum number of retries before aborting the current transaction.
      * The initial attempt is not counted as a retry.
      *
-     * If max_retries is exceeded, the transaction will return error AEROSPIKE_ERR_TIMEOUT.
+     * If <code>maxRetries</code> is exceeded, the transaction will return
+     * error {@link module:aerospike/status.ERR_TIMEOUT|ERR_TIMEOUT}.
      *
      * WARNING: Database writes that are not idempotent (such as "add")
      * should not be retried because the write operation may be performed
      * multiple times if the client timed out previous transaction attempts.
-     * It's important to use a distinct write policy for non-idempotent
-     * writes which sets max_retries = 0;
+     * It is important to use a distinct write policy for non-idempotent
+     * writes which sets <code>maxRetries</code> to zero.
      *
      * @type number
      * @default: 2 (initial attempt + 2 retries = 3 attempts)
