+++ lib/policies/batch_policy.js
@@ -19,36 +19,14 @@
 const BasePolicy = require('./base_policy')
 
 /**
- * A policy affecting the behavior of operate operations.
+ * A policy affecting the behavior of batch operations.
  *
  * @extends BasePolicy
- * @see {@link module:aerospike.policy} for supported policy values.
  */
-class OperatePolicy extends BasePolicy {
+class BatchPolicy extends BasePolicy {
   constructor (props) {
     props = props || {}
     super(props)
-
-    /**
-     * Specifies the behavior for the key.
-     *
-     * @type number
-     */
-    this.key = props.key
-
-    /**
-     * Specifies the behavior for the generation value.
-     *
-     * @type number
-     */
-    this.gen = props.gen
-
-    /**
-     * Specifies the replica to be consulted for the read operation.
-     *
-     * @type number
-     */
-    this.replica = props.replica
 
     /**
      * Specifies the number of replicas consulted when reading for the desired
@@ -56,28 +34,30 @@ class OperatePolicy extends BasePolicy {
      *
      * @type number
      * @default Aerospike.policy.consistencyLevel.ONE
+     * @see {@link module:aerospike/policy.consistencyLevel} for supported policy values.
      */
     this.consistencyLevel = props.consistencyLevel
 
     /**
-     * Specifies the number of replicas required to be committed successfully
-     * when writing before returning transaction succeeded.
-     *
-     * @type number
-     */
-    this.commitLevel = props.commitLevel
-
-    /**
-     * Specifies whether a {@link
-     * http://www.aerospike.com/docs/guide/durable_deletes.html|tombstone}
-     * should be written in place of a record that gets deleted as a result of
-     * this operation.
+     * Allow batch to be processed immediately in the server's receiving thread
+     * when the server deems it to be appropriate. If false, the batch will
+     * always be processed in separate transaction threads.
      *
      * @type boolean
-     * @default false (do not tombstone deleted records)
+     * @default <code>true</code>
      */
-    this.durableDelete = props.durableDelete
+    this.allowInline = props.allowInline
+
+    /**
+     * Send set name field to server for every key in the batch. This is only
+     * necessary when authentication is enabled and security roles are defined
+     * on a per-set basis.
+     *
+     * @type boolean
+     * @default <code>false</code>
+     */
+    this.sendSetName = props.sendSetName
   }
 }
 
-module.exports = OperatePolicy
+module.exports = BatchPolicy
