+++ lib/policies/operate_policy.js
@@ -19,27 +19,20 @@
 const BasePolicy = require('./base_policy')
 
 /**
- * A policy affecting the behavior of remove operations.
+ * A policy affecting the behavior of operate operations.
  *
  * @extends BasePolicy
- * @see {@link module:aerospike.policy} for supported policy values.
  */
-class RemovePolicy extends BasePolicy {
+class OperatePolicy extends BasePolicy {
   constructor (props) {
     props = props || {}
     super(props)
 
     /**
-     * The generation of the record.
-     *
-     * @type number
-     */
-    this.generation = props.generation
-
-    /**
      * Specifies the behavior for the key.
      *
      * @type number
+     * @see {@link module:aerospike/policy.key} for supported policy values.
      */
     this.key = props.key
 
@@ -47,14 +40,34 @@ class RemovePolicy extends BasePolicy {
      * Specifies the behavior for the generation value.
      *
      * @type number
+     * @see {@link module:aerospike/policy.gen} for supported policy values.
      */
     this.gen = props.gen
+
+    /**
+     * Specifies the replica to be consulted for the read operation.
+     *
+     * @type number
+     * @see {@link module:aerospike/policy.replica} for supported policy values.
+     */
+    this.replica = props.replica
+
+    /**
+     * Specifies the number of replicas consulted when reading for the desired
+     * consistency guarantee.
+     *
+     * @type number
+     * @default Aerospike.policy.consistencyLevel.ONE
+     * @see {@link module:aerospike/policy.consistencyLevel} for supported policy values.
+     */
+    this.consistencyLevel = props.consistencyLevel
 
     /**
      * Specifies the number of replicas required to be committed successfully
      * when writing before returning transaction succeeded.
      *
      * @type number
+     * @see {@link module:aerospike/policy.commitLevel} for supported policy values.
      */
     this.commitLevel = props.commitLevel
 
@@ -65,10 +78,10 @@ class RemovePolicy extends BasePolicy {
      * this operation.
      *
      * @type boolean
-     * @default false (do not tombstone deleted records)
+     * @default <code>false</code> (do not tombstone deleted records)
      */
     this.durableDelete = props.durableDelete
   }
 }
 
-module.exports = RemovePolicy
+module.exports = OperatePolicy
