+++ lib/policies/query_policy.js
@@ -22,7 +22,6 @@ const BasePolicy = require('./base_policy')
  * A policy affecting the behavior of query operations.
  *
  * @extends BasePolicy
- * @see {@link module:aerospike.policy} for supported policy values.
  */
 class QueryPolicy extends BasePolicy {
   constructor (props) {
