+++ lib/policies/read_policy.js
@@ -19,15 +19,30 @@
 const BasePolicy = require('./base_policy')
 
 /**
- * A policy affecting the behavior of batch operations.
+ * A policy affecting the behavior of read operations.
  *
  * @extends BasePolicy
- * @see {@link module:aerospike.policy} for supported policy values.
  */
-class BatchPolicy extends BasePolicy {
+class ReadPolicy extends BasePolicy {
   constructor (props) {
     props = props || {}
     super(props)
+
+    /**
+     * Specifies the behavior for the key.
+     *
+     * @type number
+     * @see {@link module:aerospike/policy.key} for supported policy values.
+     */
+    this.key = props.key
+
+    /**
+     * Specifies the replica to be consulted for the read operation.
+     *
+     * @type number
+     * @see {@link module:aerospike/policy.replica} for supported policy values.
+     */
+    this.replica = props.replica
 
     /**
      * Specifies the number of replicas consulted when reading for the desired
@@ -35,29 +50,10 @@ class BatchPolicy extends BasePolicy {
      *
      * @type number
      * @default Aerospike.policy.consistencyLevel.ONE
+     * @see {@link module:aerospike/policy.consistencyLevel} for supported policy values.
      */
     this.consistencyLevel = props.consistencyLevel
-
-    /**
-     * Allow batch to be processed immediately in the server's receiving thread
-     * when the server deems it to be appropriate. If false, the batch will
-     * always be processed in separate transaction threads.
-     *
-     * @type boolean
-     * @default true
-     */
-    this.allowInline = props.allowInline
-
-    /**
-     * Send set name field to server for every key in the batch. This is only
-     * necessary when authentication is enabled and security roles are defined
-     * on a per-set basis.
-     *
-     * @type boolean
-     * @default false
-     */
-    this.sendSetName = props.sendSetName
   }
 }
 
-module.exports = BatchPolicy
+module.exports = ReadPolicy
