+++ lib/policies/remove_policy.js
@@ -19,39 +19,58 @@
 const BasePolicy = require('./base_policy')
 
 /**
- * A policy affecting the behavior of read operations.
+ * A policy affecting the behavior of remove operations.
  *
  * @extends BasePolicy
- * @see {@link module:aerospike.policy} for supported policy values.
  */
-class ReadPolicy extends BasePolicy {
+class RemovePolicy extends BasePolicy {
   constructor (props) {
     props = props || {}
     super(props)
 
     /**
+     * The generation of the record.
+     *
+     * @type number
+     */
+    this.generation = props.generation
+
+    /**
      * Specifies the behavior for the key.
      *
      * @type number
+     * @see {@link module:aerospike/policy.key} for supported policy values.
      */
     this.key = props.key
 
     /**
-     * Specifies the replica to be consulted for the read operation.
+     * Specifies the behavior for the generation value.
      *
      * @type number
+     * @see {@link module:aerospike/policy.gen} for supported policy values.
      */
-    this.replica = props.replica
+    this.gen = props.gen
 
     /**
-     * Specifies the number of replicas consulted when reading for the desired
-     * consistency guarantee.
+     * Specifies the number of replicas required to be committed successfully
+     * when writing before returning transaction succeeded.
      *
      * @type number
-     * @default Aerospike.policy.consistencyLevel.ONE
+     * @see {@link module:aerospike/policy.commitLevel} for supported policy values.
      */
-    this.consistencyLevel = props.consistencyLevel
+    this.commitLevel = props.commitLevel
+
+    /**
+     * Specifies whether a {@link
+     * http://www.aerospike.com/docs/guide/durable_deletes.html|tombstone}
+     * should be written in place of a record that gets deleted as a result of
+     * this operation.
+     *
+     * @type boolean
+     * @default <code>false</code> (do not tombstone deleted records)
+     */
+    this.durableDelete = props.durableDelete
   }
 }
 
-module.exports = ReadPolicy
+module.exports = RemovePolicy
