+++ lib/policies/scan_policy.js
@@ -19,44 +19,22 @@
 const BasePolicy = require('./base_policy')
 
 /**
- * A policy affecting the behavior of apply operations.
+ * A policy affecting the behavior of scan operations.
  *
- * @extends ApplyPolicy
- * @see {@link module:aerospike.policy} for supported policy values.
+ * @extends BasePolicy
  */
-class ApplyPolicy extends BasePolicy {
+class ScanPolicy extends BasePolicy {
   constructor (props) {
     props = props || {}
     super(props)
 
     /**
-     * Specifies the behavior for the generation value.
+     * Abort the scan if the cluster is not in a stable state.
      *
-     * @type number
+     * @type boolean
+     * @default <code>false</code>
      */
-    this.gen = props.gen
-
-    /**
-     * Specifies the behavior for the key.
-     *
-     * @type number
-     */
-    this.key = props.key
-
-    /**
-     * Specifies the number of replicas required to be committed successfully
-     * when writing before returning transaction succeeded.
-     *
-     * @type number
-     */
-    this.commitLevel = props.commitLevel
-
-    /**
-     * The time-to-live (expiration) of the record in seconds.
-     *
-     * @type number
-     */
-    this.ttl = props.ttl
+    this.failOnClusterChange = props.failOnClusterChange
 
     /**
      * Specifies whether a {@link
@@ -65,10 +43,10 @@ class ApplyPolicy extends BasePolicy {
      * this operation.
      *
      * @type boolean
-     * @default false (do not tombstone deleted records)
+     * @default <code>false</code> (do not tombstone deleted records)
      */
     this.durableDelete = props.durableDelete
   }
 }
 
-module.exports = ApplyPolicy
+module.exports = ScanPolicy
