+++ lib/policies/write_policy.js
@@ -22,7 +22,6 @@ const BasePolicy = require('./base_policy')
  * A policy affecting the behavior of write operations.
  *
  * @extends BasePolicy
- * @see {@link module:aerospike.policy} for supported policy values.
  */
 class WritePolicy extends BasePolicy {
   constructor (props) {
@@ -41,6 +40,7 @@ class WritePolicy extends BasePolicy {
      * Specifies the behavior for the key.
      *
      * @type number
+     * @see {@link module:aerospike/policy.key} for supported policy values.
      */
     this.key = props.key
 
@@ -48,6 +48,7 @@ class WritePolicy extends BasePolicy {
      * Specifies the behavior for the generation value.
      *
      * @type number
+     * @see {@link module:aerospike/policy.gen} for supported policy values.
      */
     this.gen = props.gen
 
@@ -55,6 +56,7 @@ class WritePolicy extends BasePolicy {
      * Specifies the behavior for the existence of the record.
      *
      * @type number
+     * @see {@link module:aerospike/policy.exists} for supported policy values.
      */
     this.exists = props.exists
 
@@ -63,6 +65,7 @@ class WritePolicy extends BasePolicy {
      * when writing before returning transaction succeeded.
      *
      * @type number
+     * @see {@link module:aerospike/policy.commitLevel} for supported policy values.
      */
     this.commitLevel = props.commitLevel
 
@@ -73,7 +76,7 @@ class WritePolicy extends BasePolicy {
      * this operation.
      *
      * @type boolean
-     * @default false (do not tombstone deleted records)
+     * @default <code>false</code> (do not tombstone deleted records)
      */
     this.durableDelete = props.durableDelete
   }
