+++ lib/policy.js
@@ -18,15 +18,15 @@
 
 const as = require('bindings')('aerospike.node')
 
-let ApplyPolicy = require('./policy/apply_policy')
-let BatchPolicy = require('./policy/batch_policy')
-let InfoPolicy = require('./policy/info_policy')
-let OperatePolicy = require('./policy/operate_policy')
-let QueryPolicy = require('./policy/query_policy')
-let ReadPolicy = require('./policy/read_policy')
-let RemovePolicy = require('./policy/remove_policy')
-let ScanPolicy = require('./policy/scan_policy')
-let WritePolicy = require('./policy/write_policy')
+let ApplyPolicy = require('./policies/apply_policy')
+let BatchPolicy = require('./policies/batch_policy')
+let InfoPolicy = require('./policies/info_policy')
+let OperatePolicy = require('./policies/operate_policy')
+let QueryPolicy = require('./policies/query_policy')
+let ReadPolicy = require('./policies/read_policy')
+let RemovePolicy = require('./policies/remove_policy')
+let ScanPolicy = require('./policies/scan_policy')
+let WritePolicy = require('./policies/write_policy')
 
 /**
  * @module aerospike/policy
